+++ src/gitbook_downloader/gui/web/assets/swimlanesDiagram-VR7AAH4N-g8Jc88O4.js
@@ -1,8 +1,0 @@
-import{c as e,s as r}from"./flowDiagram-HODETNUW-BE7onCdn.js";import{_ as a}from"./index-CLIDcjUx.js";import"./chunk-5VM5RSS4-BTvzNq5H.js";import"./chunk-XXDRQBXY-C4KR5Pea.js";import"./chunk-POPQ4Y6H-CXal0oMS.js";import"./chunk-F27PBJKO-5nqFWuCV.js";import"./channel-C_9Q20lo.js";var o=a(t=>`${r(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,f=e({defaultLayout:"swimlane",styles:s});export{f as diagram};